// src/modules/manageOrganization/modules/workPositions/enums/infoData.ts
export enum StoreInfoKeys {
  code = 'Centro',
  denomination = 'Sucursal',
  status = 'Estatus',
  address = 'Dirección',
  zipcode = 'Código postal',
  georefDistance = 'Geocerca',
  latitude = 'Latitud',
  longitude = 'Longitud',
}

export enum CostCenterInfoKeys {
  status = 'Estatus',
  code = 'Código',
  country = 'País',
  state = 'Estado',
  city = 'Ciudad',
}

export enum CompTabInfoKeys {
  status = 'Estatus',
  minAuthorizedSalary = 'Salario Mínimo Autorizado',
  maxAuthorizedSalary = 'Salario Máximo Autorizado',
  minSalary = 'Salario Mínimo  General',
}

export enum HierarchyInfoKeys {
  orgManager = 'Cargo Superior Organigrama',
  approvalManager = 'Cargo Superior en Incidencias',
}

export enum WorkPositionCategoryInfoKeys {
  code = 'Código',
  status = 'Estatus',
}

export enum CompCategoryInfoKeys {
  status = 'Estatus',
}